misc: inv_mpu: yas530.c: Fix noisy performance.
The issue was a variable that was suppose to be unsigned was signed.
This lead to all negative numbers on the trim factors being way off.
One of the temperature compensation values was suppose to be slightly
negative and the sign error made it a large negative which caused the problem.

Change-Id: I80de9327504499e57547d9209dc06f5fdcf56636
Signed-off-by: Nathan Royer <nroyer@invensense.com>