modem_if: Drop HDLC frame if it is too large
For IP traffic, if the HDLC frame is too large, there is no point to
fragment the packet, the IP layer will drop the packet because the cecksum
is wrong. Instead, drop the packet. This should only happen if the HDLC
frame has been corruted, given the MTU. This avoids a kernel panic because
there is no space to add the eth header in the skb.

Change-Id: I35a5c8334c583f90de1e78c09ac8d4564b276a17
Signed-off-by: Benoit Goby <benoit@android.com>