OMAP4: DPLL cascading: Removed race between SR calibration and DPLL cascading exit
Before OMAP starts to exit from DPLL cascading the omap_sr_disable() is called
and when it completely exits from DPLL cascading the omap_sr_enable() is called.
A situation can occur when the OPP first calibration or periodic recalibration
of a smartreflex is continue the exit from DPLL cascading is started that stops
the calibration work by calling omap_sr_disable(). The calibration will start
again when we call omap_sr_enable(). A configuration of error generator during
the calibration is depend on flag omap4_is_in_dpll_cascading. So error generator
will be configured supposing that device is in DPLL cascading but actually it is
not. So flag indicating DPLL cascading mode has to be cleared prior to the
omap_sr_enable(), that is before the potential calibration.

Race illustration:

Assume that smartreflex calibration is ongoing and device is in DPLL cascading,
that is omap4_is_in_dpll_cascading = 1
-------------------------------------------------------------------------------
...
omap4_dpll_low_power_cascade_exit() {
  ...
  omap_sr_disable() // disable smartreflex calibration, so it will start
                    // again when we call omap_sr_enable()

  ...               // completely exit form DPLL cascading
  omap_sr_enable() {
    ...
    if (omap4_is_in_dpll_cascading()) // flag is set but cascading is off
      sys_clk_speed = 12288000;
    else
      sys_clk_speed = clk_get_rate(sys_ck);
    ...
  }
  omap4_is_in_dpll_cascading = 0
}

Change-Id: Iac03021ddc8517a09545f158865ee0d354852441
Signed-off-by: Ivan Khoronzhuk <ivan.khoronzhuk@ti.com>