mmc: omap-hsmmc: Set HS200 via DT property
The CAPA2 register (bit 1: SDR104) was used to determine
if the controller supported HS200.  However, the bit is
set if the controller supports either SDR104 or HS200, so
it's not to use it to differentiate which feature is
supported. A DT property is used instead, specifically for
the HS200 support.

Change-Id: I54bcf659a083ec0b968c80273807485ff601a049
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>