tuna-compatible version of SMC PA protocol.
The only publicly-available version of SMC PA firmware for tuna uses
older communication protocol between the "normal world" and "secure world".
This change downgrades the communication parts of the protocol without
touching the other parts of the stack. While in theory this could lead
to the incompatibilities in the other parts of the code due to versions
mismatch, things seem to work fine in my limited testing + this keeps
in place bug-fixes that are present in later version of the code + there
are less things to modify compared to the downgrade of the complete
SMC-related code base.