panel-ltn: don't try to stop the gp timer twice
When recent Android versions turn off the screen, they do so by fading the
backlight first - by slowly reducing the brightness down to 0 before actually
turning it off.
This causes a minor problem with this driver: when the brightness is set to 0,
the backlight pwm timer is stopped. However, the panel power off function tries
to stop it again, resulting in a kernel warning.

This commit modifies the power off sequence to only try stopping the timer
if it's actually running (ie. the brightness is non-zero).